Now modbus can successfully change the units of temperature

+++ modbus_device.h
@@ -44,7 +44,7 @@ public:
      * @param value The value to write
      * @return 0 if success, -1 otherwise
      */
-    int write_to_register(int location, int value);
+    int write_to_register(int location, uint16_t value);
 
     /**
      * Reads a sequential list of registers starting at the desired address. The number of registers can be
